WIP restructure
going to try out something different
