+++ bindings/wasm/test-web/html/index.html
@@ -2,7 +2,7 @@
 <html>
   <body>
     <script type="module">
-      import { VFSInterface } from './src/opfs-interface.js';
+      import { VFSInterface } from './packages/web/dist/opfs-interface.js';
       window.VFSInterface = VFSInterface;
     </script>
   </body>
